[ref]update components files
