+++ src/components/Slider/silder.css
@@ -119,7 +119,7 @@
 .right-buttom-title{
     width: 100px;
     height: 100px;
-    background-image: url(../../static/img/code.png);
+    background-image: url(../../../static/img/code.png);
     background-size: 100%,100%;
     position: relative;
     top: -20px;
