+++ src/components/SliderBugPhone/index.js
