+++ src/components/SliderBugPhone/slider-bug-phone.css
@@ -1,10 +1,10 @@
 @font-face {
     font-family: content;
-    src: url('../../static/975MaruSC-Regular.ttf');
+    src: url('../../../static/975MaruSC-Regular.ttf');
 }
 @font-face {
     font-family: title;
-    src: url('../../static/LianMengQiYiLuShuaiZhengRuiHeiTi-2.ttf');
+    src: url('../../../static/LianMengQiYiLuShuaiZhengRuiHeiTi-2.ttf');
 }
 .index-bug-p{
     width: 80%;
