+++ src/components/SliderPhone/slider-phone.css
@@ -42,7 +42,7 @@
 .phone-QRcode{
     width: 100px;
     height: 100px;
-    background-image: url(../../static/img/code.png);
+    background-image: url(../../../static/img/code.png);
     background-size: 100%,100%;
     display: inline-block;
     margin-left: 28px;
